clients: move Dashboard/Issues components
