+++ clients/apps/web/src/components/Issues/UpfrontRewards.tsx
