+++ clients/apps/web/src/stories/BadgePromotionModal.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
 import { PolarQueryClientProvider } from '@/app/providers'
-import { BadgePromotionModal } from '@/components/Dashboard/IssuePromotionModal'
+import { BadgePromotionModal } from '@/components/Issues/IssuePromotionModal'
 import { issue, user } from 'polarkit/testdata'
 
 type Story = StoryObj<typeof BadgePromotionModal>
